+++ src/modules/sound/SoundData.h
@@ -43,7 +43,7 @@ public:
 
 	// Implements Data.
 	void *getData() const;
-	int getSize() const;
+	size_t getSize() const;
 
 	virtual int getChannels() const;
 	virtual int getBitDepth() const;
@@ -59,8 +59,8 @@ private:
 
 	void load(int samples, int sampleRate, int bitDepth, int channels, void *newData = 0);
 
-	int8 *data;
-	int size;
+	uint8 *data;
+	size_t size;
 
 	int sampleRate;
 	int bitDepth;
